[Makefile] refactor fontbakery checks
moves char set checks to a new profile, adds checks to partial VF instances
